Client: fixed voice connection not resolving

+++ ytdlp.go
@@ -714,8 +714,7 @@ type YTDLPOut struct {
 			Name string `json:"name"`
 		} `json:"zu"`
 	} `json:"automatic_captions"`
-	Subtitles struct {
-	} `json:"subtitles"`
+	Subtitles            struct{}    `json:"subtitles"`
 	CommentCount         int         `json:"comment_count"`
 	Chapters             interface{} `json:"chapters"`
 	LikeCount            int         `json:"like_count"`
@@ -744,7 +743,7 @@ type YTDLPOut struct {
 			FormatID           string      `json:"format_id"`
 			FormatNote         string      `json:"format_note"`
 			SourcePreference   int         `json:"source_preference"`
-			Fps                int         `json:"fps"`
+			Fps                float64     `json:"fps"`
 			AudioChannels      interface{} `json:"audio_channels"`
 			Height             int         `json:"height"`
 			Quality            float64     `json:"quality"`
@@ -788,7 +787,7 @@ type YTDLPOut struct {
 		Width                int     `json:"width"`
 		Height               int     `json:"height"`
 		Resolution           string  `json:"resolution"`
-		Fps                  int     `json:"fps"`
+		Fps                  float64 `json:"fps"`
 		DynamicRange         string  `json:"dynamic_range"`
 		Vcodec               string  `json:"vcodec"`
 		Vbr                  float64 `json:"vbr"`
@@ -807,7 +806,7 @@ type YTDLPOut struct {
 		FormatID           string      `json:"format_id"`
 		FormatNote         string      `json:"format_note"`
 		SourcePreference   int         `json:"source_preference"`
-		Fps                int         `json:"fps"`
+		Fps                float64     `json:"fps"`
 		AudioChannels      interface{} `json:"audio_channels"`
 		Height             int         `json:"height"`
 		Quality            float64     `json:"quality"`
@@ -852,7 +851,7 @@ type YTDLPOut struct {
 	Width          int         `json:"width"`
 	Height         int         `json:"height"`
 	Resolution     string      `json:"resolution"`
-	Fps            int         `json:"fps"`
+	Fps            float64     `json:"fps"`
 	DynamicRange   string      `json:"dynamic_range"`
 	Vcodec         string      `json:"vcodec"`
 	Vbr            float64     `json:"vbr"`
